refactor: move power shaping params to power_shaping.go
