+++ x/ccv/provider/keeper/permissionless.go
@@ -2,14 +2,11 @@ package keeper
 
 import (
 	"encoding/binary"
-	"errors"
 	"fmt"
 	"strconv"
 
-	errorsmod "cosmossdk.io/errors"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/interchain-security/v5/x/ccv/provider/types"
-	ccvtypes "github.com/cosmos/interchain-security/v5/x/ccv/types"
 )
 
 // setConsumerId sets the provided consumerId
@@ -144,64 +141,6 @@ func (k Keeper) SetConsumerInitializationParameters(ctx sdk.Context, consumerId 
 func (k Keeper) DeleteConsumerInitializationParameters(ctx sdk.Context, consumerId string) {
 	store := ctx.KVStore(k.storeKey)
 	store.Delete(types.ConsumerIdToInitializationParametersKey(consumerId))
-}
-
-// GetConsumerPowerShapingParameters returns the power-shaping parameters associated with this consumer id
-func (k Keeper) GetConsumerPowerShapingParameters(ctx sdk.Context, consumerId string) (types.PowerShapingParameters, error) {
-	store := ctx.KVStore(k.storeKey)
-	bz := store.Get(types.ConsumerIdToPowerShapingParametersKey(consumerId))
-	if bz == nil {
-		return types.PowerShapingParameters{}, errorsmod.Wrapf(ccvtypes.ErrStoreKeyNotFound,
-			"GetConsumerPowerShapingParameters, consumerId(%s)", consumerId)
-	}
-	var record types.PowerShapingParameters
-	if err := record.Unmarshal(bz); err != nil {
-		return types.PowerShapingParameters{}, errorsmod.Wrapf(ccvtypes.ErrStoreUnmarshal,
-			"GetConsumerPowerShapingParameters, consumerId(%s): %s", consumerId, err.Error())
-	}
-	return record, nil
-}
-
-// SetConsumerPowerShapingParameters sets the power-shaping parameters associated with this consumer id.
-// Note that it also updates the allowlist and denylist indexes if they are different
-func (k Keeper) SetConsumerPowerShapingParameters(ctx sdk.Context, consumerId string, parameters types.PowerShapingParameters) error {
-	store := ctx.KVStore(k.storeKey)
-	bz, err := parameters.Marshal()
-	if err != nil {
-		return fmt.Errorf("failed to marshal power-shaping parameters (%+v) for consumer id (%s): %w", parameters, consumerId, err)
-	}
-
-	// get old power shaping params
-	oldParameters, err := k.GetConsumerPowerShapingParameters(ctx, consumerId)
-	// ignore ErrStoreKeyNotFound errors as this might be the first time the power shaping params are set
-	if errors.Is(err, ccvtypes.ErrStoreUnmarshal) {
-		return fmt.Errorf("cannot get consumer previous power shaping parameters: %w", err)
-	}
-
-	store.Set(types.ConsumerIdToPowerShapingParametersKey(consumerId), bz)
-
-	// update allowlist and denylist indexes if needed
-	if !equalStringSlices(oldParameters.Allowlist, parameters.Allowlist) {
-		k.UpdateAllowlist(ctx, consumerId, parameters.Allowlist)
-	}
-	if !equalStringSlices(oldParameters.Denylist, parameters.Denylist) {
-		k.UpdateDenylist(ctx, consumerId, parameters.Denylist)
-	}
-
-	return nil
-}
-
-// equalStringSlices returns true if two string slices are equal
-func equalStringSlices(a, b []string) bool {
-	if len(a) != len(b) {
-		return false
-	}
-	for i, v := range a {
-		if v != b[i] {
-			return false
-		}
-	}
-	return true
 }
 
 // GetConsumerPhase returns the phase associated with this consumer id
